+++ hist2query/app/app.py
@@ -4,7 +4,6 @@ import os
 import asyncio
 from collections.abc import Callable
 from contextlib import asynccontextmanager
-from PIL import Image
 from fastapi import FastAPI, UploadFile, File, Depends, HTTPException
 import torch
 import timm
@@ -20,7 +19,9 @@ from hist2query.app.utils import (
     decode_patch,
     load_hf_model,
     load_prism2_processing,
-    prism2_prompt_type, Prism2ChatRequestParams)
+    prism2_prompt_type,
+    Prism2ChatRequestParams,
+    extract_virchow2_embeddings)
 
 def create_app(model_loader: Callable[[], Tuple[
                timm.models.vision_transformer.VisionTransformer,
@@ -76,8 +77,9 @@ def create_app(model_loader: Callable[[], Tuple[
                                        "was not supplied to the hist2query deployment.")
 
         tile_rgb = await decode_patch(patch)
+        tile_rgb = make_tiles(tile_rgb)
 
-        batch = preprocess_tiles(make_tiles(tile_rgb))
+        batch = await asyncio.to_thread(preprocess_tiles,tile_rgb)
 
         async with app.state.inference_lock:
             # a single transform per image is faster, but results seem notieceably worse
@@ -128,22 +130,18 @@ def create_app(model_loader: Callable[[], Tuple[
 
         tile_rgb = make_tiles(tile_rgb)
 
-        tile_batch = []
-        # cannot use the same tile generation function as for uni2 as only the class token is pulled from Virchow2
-        for tile_rgb in tile_rgb:
-            output = app.state.virchow2_model(app.state.virchow2_transform(
-                Image.fromarray(tile_rgb).convert("RGB")).unsqueeze(0))
-            tile_batch.append(output[:, 0])
-
-        tile_batch = torch.cat(tile_batch, dim=0)
-        batch = app.state.prism2_processor([tile_batch])
-        async with app.state.inference_lock:
-            batch = batch.to(app.state.device)
-            with torch.autocast(app.state.device, torch.bfloat16):
-                answers = prism2_prompt_type(app.state.prism2_model, str(params.question),
-                                         batch, int(params.max_token_response),
-                                         params.raw_scores_binary, params.binary_threshold_for_yes)
-        del tile_rgb, tile_batch, batch
-        return {'response': answers}
+        tile_batch = await asyncio.to_thread(extract_virchow2_embeddings,
+                app.state.virchow2_model, app.state.virchow2_transform, tile_rgb)
+        if tile_batch is not None:
+            batch = app.state.prism2_processor([tile_batch])
+            async with app.state.inference_lock:
+                batch = batch.to(app.state.device)
+                with torch.autocast(app.state.device, torch.bfloat16):
+                    answers = prism2_prompt_type(app.state.prism2_model, str(params.question),
+                                                 batch, int(params.max_token_response),
+                                                 params.raw_scores_binary, params.binary_threshold_for_yes)
+            del tile_rgb, tile_batch, batch
+            return {'response': answers}
+        return {'response': 'Error: no tiles computed.'}
 
     return app