+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Move embedding to device casting for GPU inside `asyncio.Lock()`
+- Put tile processing on `asyncio` async threading
 - Ensure L2 normalization match for embedding addition
 
 ## [0.3.0] - 2026-08-24